+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.1.0] - 2022-08-11
+
+feat(refactor interface):
+- remove storeName, defaultValue from `init` signature
 ## [1.0.0] - 2022-08-04
 
 feat(initial):